competitions/table/league_of_legends_winning_prediction/pyproject.toml
[project]
name = "league-of-legends-winning-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lightgbm>=4.6.0",
    "marimo>=0.17.0",
    "matplotlib>=3.10.7",
    "nbformat>=5.10.4",
    "numpy>=2.3.4",
    "optuna>=4.5.0",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
]
